Limit cert_selectors source files to one logger ID
